# Cut parse time almost in half