@dtolnay 2018-09-04
Serde and serde_json author here by the way. :stuck_out_tongue_winking_eye: Feel free to just close the PR if the new code is too crazy. I know I used one or two Rust features that were not present in the code before.