@bcantrill 2018-09-04
This is awesome -- thank you!  For whatever it's worth, here is [the exact input file I used in my demo](https://us-east.manta.joyent.com/bcantrill/public/statemap/pg-zfs.out.gz) (~30MB compressed). As I mentioned there, I really haven't done *any* analysis on this, so it's not a surprise that there is some low-hanging fruit. (And I definitely missed serde_json's ability to handle concatenated JSON!)  That said, I would love to know the methodology here: if you don't mind me asking, what tools do you/did you use for profiling?  Thanks again for both serde and serde_json (truly awesome work) -- and for your interest in my little goober!