@dtolnay 2018-09-04
Just callgrind in this case.

```console
$ valgrind --tool=callgrind ./target/release/statemap sample.out >rust.svg
$ callgrind_annotate --auto=yes callgrind.out.*
```

Here is the top of the callgrind output before this PR. (Again isolated to parsing only.)

```console
3,532,171,764  PROGRAM TOTALS

1,093,993,901  statemap::statemap::Statemap::json_end
  486,996,711  <serde_json::read::StrRead<'a> as serde_json::Read<'a>>::parse_str
  458,993,115  <&'a mut serde_json::de::Deserializer<R> as serde::de::Deserializer<'de>>::deserialize_struct
  138,999,100  statemap::statemap::Statemap::ingest
  111,998,799  <&'a mut serde_json::de::Deserializer<R> as serde::de::Deserializer<'de>>::deserialize_string
   90,998,635  core::num::<impl core::str::FromStr for u64>::from_str
   90,002,430  jemalloc.c:mallocx
   72,001,944  jemalloc.c:sdallocx
   65,896,974  jemalloc.c:isfree
   60,158,832  tcache.h:isfree
   60,071,035  tcache.h:mallocx
   56,999,145  serde_json::de::from_str
   51,999,220  <core::marker::PhantomData<T> as serde::de::DeserializeSeed<'de>>::deserialize
   49,999,328  statemap::statemap::Statemap::json_start
```

From this I know:

- 30% of instruction reads are in `json_end`. The `json_end` and `json_start` aren't even necessary so if they show up this high it warrants a different approach. **\*can fix\***
- The `parse_str` is serde_json tokenizing JSON strings. Makes sense, practically the entire input line consists of strings. This should be the bulk of the actual useful work.
- The output of deserialization consists almost entirely of structs and strings so `deserialize_struct` and `deserialize_string` are not surprising.
- `u64::from_str` is slow. If your numbers weren't quoted, serde_json would handle this without using `u64::from_str` and be faster. I could change your dtrace scripts but meh.
- There really shouldn't be any malloc in this workload. Serde_json internally isn't mallocing unless you have strings containing escape sequences, so it's only showing up because the output object holds a `String` which forces an allocation and memcpy. **\*can fix\***

Then I confirm that what callgrind suggests likely to help does actually help (it doesn't always). The old-fashioned way:

```
$ time ./target/release/statemap sample.out >rust.svg
```