@bcantrill 2018-09-18
Just wanted to follow up over here to let you know that I haven't dropped this! I've been looking at the performance of your suggestions (both of which improve performance, but I'm trying to quantify them separately and in a lab environment), and taking apart why Rust is performing so much better than C. ;) Anyway, more to come -- and thank you again for your time and help!