@gnzlbg 2018-09-19
@dtolnay 

> The old-fashioned way:
>
> ```
> $ time ./target/release/statemap sample.out >rust.svg
> ```

You are probably already aware of this but @bcantrill might not be. There is a neatly useful tool for benchmarking cli applications called [`hyperfine`](https://github.com/sharkdp/hyperfine) that's available in some package repositories (`brew install hyperfine`). Its written in Rust, so a `cargo install hyperfine` will do as well. I found it a bit more reliable than `time x` when benchmarking cli programs recently (you can do a warm up, it measures shell start time, you can control how often the cli app gets executed, etc.).