@bcantrill 2018-09-22
Analysis of these improvements can be found in #42; I have integrated them under the aegis of #43 and #43, and am therefore closing this pull request. Thank you again, @dtolnay! (And thanks, too to @gnzlbg for the hyperfine pointer!)